fix: test fixes
