+++ tests/unit/tests_filman_server/database/tests_queries/test_filmweb.py
@@ -112,277 +112,6 @@ def test_db():
 
 
 #
-# USERS
-#
-
-
-def test_get_user_by_id(test_db):
-    result = crud.get_user(test_db, id=1, filmweb_id=None, discord_id=None)
-    assert result.id == 1
-    assert result.discord_id == 123456789
-
-
-def test_get_user_by_filmweb_id(test_db):
-    result = crud.get_user(test_db, id=None, filmweb_id="filmweb123", discord_id=None)
-    assert result.id == 2
-    assert result.discord_id == 987654321
-
-
-def test_get_user_by_discord_id(test_db):
-    result = crud.get_user(test_db, id=None, filmweb_id=None, discord_id=123456789)
-    assert result.id == 1
-    assert result.discord_id == 123456789
-
-
-def test_get_user_no_parameters(test_db):
-    result = crud.get_user(test_db, id=None, filmweb_id=None, discord_id=None)
-    assert result is None
-
-
-def test_get_user_by_non_existent_id(test_db):
-    result = crud.get_user(test_db, id=999, filmweb_id=None, discord_id=None)
-    assert result is None
-
-
-def test_get_user_by_non_existent_filmweb_id(test_db):
-    result = crud.get_user(test_db, id=None, filmweb_id="nonexistent", discord_id=None)
-    assert result is None
-
-
-def test_get_user_by_non_existent_discord_id(test_db):
-    result = crud.get_user(test_db, id=None, filmweb_id=None, discord_id=999999999)
-    assert result is None
-
-
-def test_get_user_empty_database():
-    engine = create_engine("sqlite:///:memory:")
-    models.Base.metadata.create_all(engine)
-    Session = sessionmaker(bind=engine)
-    session = Session()
-
-    result = crud.get_user(session, id=1, filmweb_id=None, discord_id=None)
-    assert result is None
-
-    session.close()
-    models.Base.metadata.drop_all(engine)
-
-
-#
-# DISCORD GUILDS
-#
-
-
-def test_get_guild_by_id(test_db):
-    result = crud.get_guild(test_db, discord_guild_id=123456789)
-    assert result.discord_guild_id == 123456789
-    assert result.discord_channel_id == 100000
-
-
-def test_set_guild(test_db):
-    guild = models.DiscordGuilds(discord_guild_id=444444444, discord_channel_id=600000)
-    result = crud.set_guild(test_db, guild)
-    assert result.discord_guild_id == 444444444
-    assert result.discord_channel_id == 600000
-
-    result = crud.get_guild(test_db, discord_guild_id=444444444)
-    assert result.discord_guild_id == 444444444
-    assert result.discord_channel_id == 600000
-
-
-def test_delete_guild(test_db):
-    result = crud.delete_guild(test_db, discord_guild_id=444444444)
-    assert result.discord_guild_id == 444444444
-
-    result = crud.get_guild(test_db, discord_guild_id=444444444)
-    assert result is None
-
-
-def test_get_guilds(test_db):
-    result = crud.get_guilds(test_db)
-    assert len(result) == 5
-
-
-def test_get_guild_members(test_db):
-    result = crud.get_guild_members(test_db, discord_guild_id=123456789)
-    assert len(result) == 3
-
-    result = crud.get_guild_members(test_db, discord_guild_id=987654321)
-    assert len(result) == 3
-
-    result = crud.get_guild_members(test_db, discord_guild_id=111111111)
-    assert len(result) == 3
-
-    result = crud.get_guild_members(test_db, discord_guild_id=222222222)
-    assert len(result) == 1
-
-    assert isinstance(result[0], models.User)
-
-
-#
-# DISCORD DESTINATIONS
-#
-
-# MULTIPLE DESTINATIONS
-
-
-def test_get_user_destinations_by_user_id(test_db):
-    # User 1 should be in 1 guild
-    result = crud.get_user_destinations(test_db, user_id=1, discord_user_id=None)
-    assert len(result) == 1
-    assert result[0].discord_guild_id == 123456789
-
-    # User 2 should be in 2 guilds
-    result = crud.get_user_destinations(test_db, user_id=2, discord_user_id=None)
-    assert len(result) == 2
-    assert result[0].discord_guild_id in [987654321, 111111111]
-
-    # User 3 should be in 3 guilds
-    result = crud.get_user_destinations(test_db, user_id=3, discord_user_id=None)
-    assert len(result) == 3
-    assert result[0].discord_guild_id in [123456789, 987654321, 111111111]
-
-    # User 4 should be in 4 guilds
-    result = crud.get_user_destinations(test_db, user_id=4, discord_user_id=None)
-    assert len(result) == 4
-    assert result[0].discord_guild_id in [123456789, 987654321, 111111111, 222222222]
-
-
-def test_get_user_destinations_by_discord_user_id(test_db):
-    # User 1 should be in 1 guild
-    result = crud.get_user_destinations(test_db, user_id=None, discord_user_id=123456789)
-    assert len(result) == 1
-    assert result[0].discord_guild_id == 123456789
-
-    # User 2 should be in 2 guilds
-    result = crud.get_user_destinations(test_db, user_id=None, discord_user_id=987654321)
-    assert len(result) == 2
-    assert result[0].discord_guild_id in [123456789, 111111111]
-
-    # User 3 should be in 3 guilds
-    result = crud.get_user_destinations(test_db, user_id=None, discord_user_id=777777777)
-    assert len(result) == 3
-    assert result[0].discord_guild_id in [123456789, 987654321, 111111111]
-
-    # User 4 should be in 4 guilds
-    result = crud.get_user_destinations(test_db, user_id=None, discord_user_id=888888888)
-    assert len(result) == 4
-    assert result[0].discord_guild_id in [123456789, 987654321, 111111111, 222222222]
-
-
-def test_get_user_destinations_by_non_existent_user_id(test_db):
-    result = crud.get_user_destinations(test_db, user_id=999, discord_user_id=None)
-    assert result is None
-
-
-def test_get_user_destinations_by_non_existent_discord_user_id(test_db):
-    result = crud.get_user_destinations(test_db, user_id=None, discord_user_id=999999999)
-    assert result is None
-
-
-# SINGLE DESTINATION
-
-
-def test_get_user_destination_by_user_id_user_in_guild(test_db):
-    result = crud.get_user_destination(test_db, user_id=1, discord_user_id=None, discord_guild_id=123456789)
-    assert result.user_id == 1
-
-    result = crud.get_user_destination(test_db, user_id=2, discord_user_id=None, discord_guild_id=987654321)
-    assert result.user_id == 2
-
-    result = crud.get_user_destination(test_db, user_id=3, discord_user_id=None, discord_guild_id=123456789)
-    assert result.user_id == 3
-
-    result = crud.get_user_destination(test_db, user_id=4, discord_user_id=None, discord_guild_id=123456789)
-    assert result.user_id == 4
-
-    result = crud.get_user_destination(test_db, user_id=4, discord_user_id=None, discord_guild_id=987654321)
-    assert result.user_id == 4
-
-
-def test_get_user_destination_by_user_id_user_not_in_guild(test_db):
-    result = crud.get_user_destination(test_db, user_id=1, discord_user_id=None, discord_guild_id=987654321)
-    assert result is None
-
-    result = crud.get_user_destination(test_db, user_id=2, discord_user_id=None, discord_guild_id=123456789)
-    assert result is None
-
-    result = crud.get_user_destination(test_db, user_id=3, discord_user_id=None, discord_guild_id=222222222)
-    assert result is None
-
-    result = crud.get_user_destination(test_db, user_id=4, discord_user_id=None, discord_guild_id=333333333)
-    assert result is None
-
-
-# SET DESTINATION
-
-
-def test_set_user_destination(test_db):
-    # let user 1 be in 5 guilds
-    result = crud.set_user_destination(test_db, user_id=1, discord_guild_id=123456789)
-    assert result.user_id == 1
-
-    result = crud.set_user_destination(test_db, user_id=1, discord_guild_id=987654321)
-    assert result.user_id == 1
-
-    result = crud.set_user_destination(test_db, user_id=1, discord_guild_id=111111111)
-    assert result.user_id == 1
-
-    result = crud.set_user_destination(test_db, user_id=1, discord_guild_id=222222222)
-    assert result.user_id == 1
-
-    result = crud.set_user_destination(test_db, user_id=1, discord_guild_id=333333333)
-    assert result.user_id == 1
-
-    # check if user 1 is in 5 guilds
-    result = crud.get_user_destinations(test_db, user_id=1, discord_user_id=None)
-    assert len(result) == 5
-
-    for id in range(2, 5):
-        result = crud.get_user_destinations(test_db, user_id=id, discord_user_id=None)
-        assert len(result) == id
-
-
-# DELETE DESTINATION
-
-
-def test_delete_user_destination(test_db):
-    # remove 3 guilds from user 1
-
-    result = crud.delete_user_destination(test_db, user_id=1, discord_user_id=None, discord_guild_id=123456789)
-    assert result.user_id == 1
-
-    result = crud.delete_user_destination(test_db, user_id=1, discord_user_id=None, discord_guild_id=987654321)
-    assert result.user_id == 1
-
-    result = crud.delete_user_destination(test_db, user_id=1, discord_user_id=None, discord_guild_id=111111111)
-    assert result.user_id == 1
-
-    result = crud.get_user_destinations(test_db, user_id=1, discord_user_id=None)
-    assert len(result) == 2
-
-    for id in range(2, 5):
-        result = crud.get_user_destinations(test_db, user_id=id, discord_user_id=None)
-        assert len(result) == id
-
-
-# DELETE DESTINATIONS
-
-
-def test_delete_user_destinations(test_db):
-    # remove all guilds from user 1
-
-    result = crud.delete_user_destinations(test_db, user_id=1, discord_user_id=None)
-    assert result == None
-
-    result = crud.get_user_destinations(test_db, user_id=1, discord_user_id=None)
-    assert result is None or len(result) == 0
-
-    for id in range(2, 5):
-        result = crud.get_user_destinations(test_db, user_id=id, discord_user_id=None)
-        assert len(result) == id
-
-
-#
 # FILMWEB MOVIES
 #
